==> notebooks/03_create_html.ipynb <==
import folium
import pandas as pd
import datautils

import numpy as np
from PIL import Image
import maputils

# create folder structures and holds constants
data_path = datautils.DataPaths("../data")

import pandas as pd
import folium
from folium.plugins import Geocoder
from folium.plugins.treelayercontrol import TreeLayerControl


m = folium.Map(
    location=[47.0707, 15.4395], zoom_start=10
)  # Coordinates for Graz, Austria


Geocoder().add_to(m)


bounds_graz = [[47.0707, 15.4395], [47.0907, 15.4595]]
bounds_styria_img = [
    [46.42536839691116, 13.10432122749762],
    [47.996278912987734, 16.61222420111168],
]


zindex_last = 0
# used_extension = ".png"
used_extension = ".webp"


def add_image_layer(
    img_path: str, name: str, bounds, m: folium.Map, show=False, opacity=0.6
):

    global zindex_last
    zindex_last += 1

    img_overlay = folium.raster_layers.ImageOverlay(
        name=name,
        image=img_path,
        bounds=bounds,
        opacity=opacity,
        interactive=True,
        cross_origin=False,
        zindex=zindex_last,
    )

    feat_group = folium.FeatureGroup(name=name, show=show)
    img_overlay.add_to(feat_group)
    feat_group.add_to(m)


houses_density = []
houses_density.append(
    {
        "name": "Gebiete mit mehr als 1000 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_1000to100000_per_km2",
        "show": True,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 500 bis 1000 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_500to1000_per_km2",
        "show": True,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 250 bis 500 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_250to500_per_km2",
        "show": True,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 100 bis 250 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_100to250_per_km2",
        "show": False,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 50 bis 100 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_50to100_per_km2",
        "show": False,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 10 bis 50 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_10to50_per_km2",
        "show": False,
    }
)
houses_density.append(
    {
        "name": "Gebiete mit 1 bis 10 Häusern im Umkreis von 0,5km",
        "image": "houses_dist_1to10_per_km2",
        "show": False,
    }
)

for density in houses_density:
    add_image_layer(
        f"{data_path.maps}/{density['image']}{used_extension}",
        density["name"],
        bounds_styria_img,
        m,
        show=density["show"],
        opacity=0.5,
    )

add_image_layer(
    f"{data_path.maps}/combined_data_masked{used_extension}",
    "Positive und negative Effekte kombiniert",
    bounds_styria_img,
    m,
    show=True,
)
# positive
add_image_layer(
    f"{data_path.maps}/positive_masked{used_extension}",
    "Positive Effekte",
    bounds_styria_img,
    m,
)

# add_image_layer(
#     f"{data_path.maps}/houses_density{used_extension}",
#     "Wohngebiete nach Anzahl der Häuser innerhalb eines Radius von 0,5km",
#     bounds_styria_img,
#     m,
# )


add_image_layer(
    f"{data_path.maps}/infrastructure_supermarkets_close_styria_masked{used_extension}",
    "Supermärkte (bis 1,5km Nahbereich, max. 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/infrastructure_supermarkets_diversity_styria_masked{used_extension}",
    "Supermärkte Auswahl (mind. 2, bis 1,5km Nahbereich, max. 10km), logaritmisch",
    bounds_styria_img,
    m,
)

add_image_layer(
    f"{data_path.maps}/infrastructure_restaurant_styria_masked{used_extension}",
    "Restaurants und Cafes (max. x logaritmisch)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/infrastructure_bakery_styria_masked{used_extension}",
    "Bäkereien (bis 6km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/infrastructure_drogerie_styria_masked{used_extension}",
    "Drogerien (bis 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/infrastructure_pharmacy_styria_masked{used_extension}",
    "Apotheken (bis 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/school_elementary_styria_masked{used_extension}",
    "Volksschulen (bis 4km Nahbereich, max. 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/school_nms_styria_masked{used_extension}",
    "Mittelschulen (bis 4km Nahbereich, max. 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/school_ahs_styria_masked{used_extension}",
    "AHS (bis 4km Nahbereich, max. 25km, max 3)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/transport_busstations_styria_masked{used_extension}",
    "Busstationen (bis 3km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/transport_trainstations_styria_masked{used_extension}",
    "Bahnhöfe/Haltestellen (bis 4km Nahbereich, max. 10km)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/transport_motorway_exits_styria_masked{used_extension}",
    "Autobahnabfahrten (bis 10km)",
    bounds_styria_img,
    m,
)


# negative

add_image_layer(
    f"{data_path.maps}/negative_masked{used_extension}",
    "Negative Effekte kombiniert (Straßen, Tankstellen, Fußballplätze)",
    bounds_styria_img,
    m,
    show=False,
)

add_image_layer(
    f"{data_path.maps}/negativ_infrastructure_gasstations_styria_masked{used_extension}",
    "Tankstellen (0.4km Umkreis)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/negativ_leisure_soccer_styria_masked{used_extension}",
    "Fußballplätze > 6.000m² (0,5km Umkreis)",
    bounds_styria_img,
    m,
)


add_image_layer(
    f"{data_path.maps}/negativ_infrastructure_motorway_styria_masked{used_extension}",
    "Autobahn/Schnellstraßen (ohne Tunnel, 2km Umkreis)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/negativ_infrastructure_street_pimary_secondary_styria_masked{used_extension}",
    "Bundesstraßen (ohne Tunnel, 100m Umkreis)",
    bounds_styria_img,
    m,
)
add_image_layer(
    f"{data_path.maps}/negativ_infrastructure_railway_rails_styria_masked{used_extension}",
    "Bahngleise (ohne Tunnel, 1km Umkreis)",
    bounds_styria_img,
    m,
)


# # Load the area.csv file
# area_df = pd.read_csv('areas.csv', index_col=0)
# area_df.columns = ['center_lat', 'center_lon', 'area']

# # Add markers to the map
# for idx, row in area_df.iterrows():
#     folium.Marker(
#         location=[row['center_lat'], row['center_lon']],
#         popup=row['area'],
#         icon=folium.Icon(color='blue', icon='info-sign')
#     ).add_to(m)


note_positive_effects = """"Kombinierte positive Eigenschaften der Region, dies beinhaltet:
* Anbindung an das öffentliche Verkehrsnetz (Busstationen, Bahnhöfe)
* Anbingung an die Autobahn
* Schulen (Volksschulen, Mittelschulen, AHS)
* Nahversorung und sonstige Einkaufsmöglichkeiten
* Restaurants und Cafes

Die Bewertung der Distanzen sowie die Gewichtung der einzelnen Punkte ist subejktiv nach meinen Bedürfnissen gewählt, kann jedoch beliebig angepasst werden.
siehe Github Repository für mehr Informationen:
github.com/TODO
"""

note_positive_short = """"Kombinierte positive Eigenschaften der Region, siehe github.com/TODO für mehr Information
"""

folium.LayerControl(collapsed=False, name="Layer Control").add_to(m)

# example for tree control structure
# overlay_tree = {
#     "label": "Points of Interest",
#     "select_all_checkbox": "Un/select all",
#     "children": [
#         {
#             "label": "Europe",
#             "select_all_checkbox": True,
#             "children": [
#                 {
#                     "label": "France",
#                     "select_all_checkbox": True,
#                     "children": [
#                         { "label": "Tour Eiffel", "layer": Marker([48.8582441, 2.2944775]).add_to(m) },
#                         { "label": "Notre Dame", "layer": Marker([48.8529540, 2.3498726]).add_to(m) },
#                         { "label": "Louvre", "layer": Marker([48.8605847, 2.3376267]).add_to(m) },
#                     ]
#                 }, {
#                     "label": "Germany",
#                     "select_all_checkbox": True,
#                     "children": [
#                         { "label": "Branderburger Tor", "layer": Marker([52.5162542, 13.3776805]).add_to(m)},
#                         { "label": "Kölner Dom", "layer": Marker([50.9413240, 6.9581201]).add_to(m)},
#                     ]
#                 }, {"label": "Spain",
#                     "select_all_checkbox": "De/seleccionar todo",
#                     "children": [
#                         { "label": "Palacio Real", "layer": Marker([40.4184145, -3.7137051]).add_to(m)},
#                         { "label": "La Alhambra", "layer": Marker([37.1767829, -3.5892795]).add_to(m)},
#                     ]
#                 }
#             ]
#         }, {
#             "label": "Asia",
#             "select_all_checkbox": True,
#             "children": [
#                 {
#                     "label": "Jordan",
#                     "select_all_checkbox": True,
#                     "children": [
#                         { "label": "Petra", "layer": Marker([30.3292215, 35.4432464]).add_to(m) },
#                         { "label": "Wadi Rum", "layer": Marker([29.6233486, 35.4390656]).add_to(m) }
#                     ]
#                 }, {
#                 }
#             ]
#         }
#     ]
# }

# TreeLayerControl(overlay_tree=overlay_tree, collapsed=False).add_to(m)

# Save the map with the image overlay
m.save(f"{data_path.html}/map_all_layers.html")
